requirements.txt
python-telegram-bot==20.6
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2